+++ lib/src/wrapper.dart
@@ -1,6 +1,6 @@
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/material.dart';
-import 'package:hidden_logo/src/parser.dart';
+import 'package:hidden_logo/src/base.dart';
 
 /// Logo builder function
 typedef LogoBuilder = Widget Function(
@@ -9,7 +9,7 @@ typedef LogoBuilder = Widget Function(
 );
 
 /// Determines when to show your logo
-enum LogoShowType {
+enum LogoVisibilityMode {
   /// Always show the logo
   always,
 
@@ -22,14 +22,14 @@ enum LogoShowType {
 /// widget built via provided functions on top of the screen where it it
 /// not visible under the physical hardware barrier.
 /// {@endtemplate}
-class HiddenLogo extends StatefulWidget {
+class HiddenLogo extends StatelessWidget {
   /// {@macro hidden_logo.HiddenLogo}
   const HiddenLogo({
     required this.body,
     required this.notchBuilder,
     required this.dynamicIslandBuilder,
-    this.showType = LogoShowType.always,
-    this.isShown = true,
+    this.visibilityMode = LogoVisibilityMode.always,
+    this.isVisible = true,
     super.key,
   });
 
@@ -45,98 +45,20 @@ class HiddenLogo extends StatefulWidget {
   final LogoBuilder dynamicIslandBuilder;
 
   /// Determines when to show your logo
-  final LogoShowType showType;
+  final LogoVisibilityMode visibilityMode;
 
   /// You can force hiding of the logo whilst the parameter is set to false
-  final bool isShown;
-
-  @override
-  State<HiddenLogo> createState() => _HiddenLogoState();
-}
-
-class _HiddenLogoState extends State<HiddenLogo> with WidgetsBindingObserver {
-  bool _isForeground = true;
-
-  @override
-  void initState() {
-    WidgetsBinding.instance.addObserver(this);
-    super.initState();
-  }
-
-  @override
-  void didChangeAppLifecycleState(AppLifecycleState state) {
-    switch (state) {
-      case AppLifecycleState.resumed:
-        _isForeground = true;
-        break;
-      case AppLifecycleState.inactive:
-      case AppLifecycleState.detached:
-      case AppLifecycleState.hidden:
-      case AppLifecycleState.paused:
-        _isForeground = false;
-        break;
-    }
-    super.didChangeAppLifecycleState(state);
-    setState(() {});
-  }
+  final bool isVisible;
 
   @override
   Widget build(BuildContext context) {
-    return OrientationBuilder(
-      builder: (context, orientation) {
-        if (orientation == Orientation.landscape) return widget.body;
-        return FutureBuilder(
-          future: DeviceInfoPlugin().deviceInfo,
-          builder: (context, snapshot) {
-            if (!snapshot.hasData ||
-                snapshot.hasError ||
-                snapshot.data == null) {
-              return widget.body;
-            }
-            final parser =
-                HiddenLogoParser(deviceInfo: snapshot.data! as BaseDeviceInfo);
-            final constraints = parser.logoConstraints;
-            return Stack(
-              children: [
-                widget.body,
-                if (parser.isTargetDevice &&
-                    widget.isShown &&
-                    (widget.showType == LogoShowType.always
-                        ? true
-                        : !_isForeground))
-                  Align(
-                    alignment: Alignment.topCenter,
-                    child: Padding(
-                      padding: EdgeInsets.only(
-                        top: parser.dynamicIslandTopMargin,
-                      ),
-                      child: parser.logoType == LogoType.notch
-                          ? widget.notchBuilder(
-                              context,
-                              constraints,
-                            )
-                          : ClipRRect(
-                              borderRadius: const BorderRadius.all(
-                                Radius.circular(100.0),
-                              ),
-                              child: widget.dynamicIslandBuilder(
-                                context,
-                                constraints,
-                              ),
-                            ),
-                    ),
-                  ),
-              ],
-            );
-          },
-        );
-      },
+    return HiddenLogoBase(
+      body: body,
+      notchBuilder: notchBuilder,
+      dynamicIslandBuilder: dynamicIslandBuilder,
+      visibilityMode: visibilityMode,
+      isVisible: isVisible,
+      deviceInfoPlugin: DeviceInfoPlugin(),
     );
-  }
-
-  @override
-  void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
-    super.dispose();
   }
 }